[kernel][arm64][efi] spiff up the EFI boot path to allocate the ramdisk and kernel wherever it wants
Instead of hard coding the address of the kernel load spot and ramdisk,
ask efi to allocate a block of memory large enough to hold the ramdisk +
kernel and copy both of them there.

This means the kernel will run at an arbitrary physical address, but
this is okay.

Also properly exit boot services

Change-Id: Ic297694bd125369d293501d690fd75cd383fbd16