[dev][usb-xhci] Adds functions for converting a TRB into a NO-OP, and mapping from a physical address back to a TRB pointer.
This is for use with the debug capability.

When the DbC clears halts, it writes the next TRB physical address
to be executed, but doesn't appear to guarantee where it'll be (XHCI Spec 7.6.4.3).
We also can't change it either while DbC is in the Configured state.

The easiest thing to do is convert the pending TRBs on the transfer
ring into no ops, and requeue pending requests again. The endpoint will
be in the stopped state and have ownership over the transfer ring.

Change-Id: I7f46909cd5ab68f4aecdef0633bf17f75ae7c224