[hid-parser] Add example fuzzer
This CL adds a simple fuzzer of the hid-parser as an example of how to
add fuzz targets to Zircon.

NOTE: make/engine.mk is broken in this change; the attempt to add only
the static libFuzzer library without and C++ or C++ ABI shared objects
causes clang to fail.

SEC-44 #comment Example of Zircon layer implementation

Change-Id: Iac402d5b6ce9dbab74bc37b6a1e9a8c88a2605c5