[dev][platform-bus] Move MMIO/IRQ allocations to proxy devhosts
Change pdev proxy drivers so that on creation they request resource
handles for their MMIOs and IRQs from the platform bus. Use these
resource handles to create the objects within the proxy devhost
itself, and set the name of the vmos for better bookkeeping.

This change integrates the PhysicalAspaceManager and exclusive
resource allocation into the ARM platform.

Change-Id: I8ea30c87a94952d0314f62a02ec6343bba37e03e