[ulib][ralloc] Add a cookie value to RegionAllocator::Region
A uint64_t cookie value has been added to the Region class. By doing it
in the Region class itself rather than the base ralloc_region_t it
allows for continued use of C style struct initializers with GetRegion.
Due to the const nature of returned regions from GetRegion, this cookie
is set immutably as an optional argument to GetRegion.  These changes
have the side effect of removing partiy between the C and C++ apis, but
only for the cookie. It is considered an acceptable cost to avoid all
callers using C style struct initialization needing to specify a default
cookie value.

The impact this change will have is an extra 8 bytes of memory per
region pulled out of an allocator. On a Pixelbook there are presently
~30 regions allocated at boot. This is miniscule compared to the benefit
the bookkeeping adds for clients of the library.

Test: region-alloc-test

Change-Id: I3f3eb65e990f94af58f71c6cc72fe36694632a13