[dev][platform-bus] Use separate classes for platform devices and protocol implementation devices
The class PlatformDevice in the platform bus driver was used for two purposes:

1) To implement platform devices, which run in separate devhosts.
   In this role the class did not need to implement the platform device protocol,
   but instead responded to proxied requests via the rxrpc callback from the PlatformProxy
   class in the other devhost.

2) To bind platform protocol implementation drivers, which run in the same devhost and do require
   a local platform device protocol implementation.

To simplify the code, the PlatformDevice class implements use case 1), and a new class called
ProtocolDevice implements use case 2.

In addition, protocol implementation drivers are now given a restricted version of the
platform bus protocol that disables pbus_device_add() and pbus_protocol_device_add().
There is no good use case for protocol implementation drivers to use these,
so lets not allow it.

TEST: manual testing on qemu, VIM2 and gauss.

Change-Id: I2cbe1bf0d47fac00f37093275eee4db450776f78