[zircon][sdio]Fix multiblock support and re-enable it
When submitting a multiblock CMD53, the config register in the
controller should be updated to reflect the new block length.
This changeset also fixes how the multiblk command is submitted
in PIO mode.

Test: Modified the block size of sdio fn0 to be 8 and read 16 bytes of data
on vim2 with both dma and pio mode. Prior to this change the read fails.
More testing is in progress

Change-Id: Iefce824d547de88bebd468f1d16c5f2177bfb194